backend/requirements.txt
asyncpg==0.30.0
crewai==1.7.1
fastapi==0.124.4
fitz==0.0.1.dev2
gender_guesser==0.4.0
joblib==1.5.2
numpy==2.3.5
openai==2.12.0
pandas==2.3.3
pinecone==8.0.0
python-dotenv==1.2.1
shap==0.50.0
SQLAlchemy==2.0.45
